requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.1
biopython==1.84
blinker==1.8.2
Bootstrap-Flask==2.4.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confsmooth==1.0.0
dnspython==2.6.1
dominate==2.9.1
email-validator==2.2.0
flask==3.0.3
flask-mail==0.10.0
flask-wtf==1.2.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
kombu==5.4.2
MarkupSafe==2.1.5
numpy==2.0.2
numpyencoder==0.3.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
plotly==5.24.1
prompt-toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.2
rapidfuzz==3.10.0
redis==5.0.8
requests==2.32.3
scipy==1.13.1
six==1.16.0
statsmodels==0.14.3
tenacity==9.0.0
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
visitor==0.1.3
wcwidth==0.2.13
werkzeug==3.0.4
whittaker-eilers==0.1.3
wtforms==3.1.2
zipp==3.20.2
kaleido==0.2.1
